Refactor CLI into modular command architecture
